[awful-ui] tidy reorganize components so they are grouped by function
