+++ sites/awful-ui/src/AwfulUI.tsx
@@ -5,19 +5,19 @@ import '@xyflow/react/dist/style.css';
 
 import useAwfulFlow from './hooks/useAwfulFlow';
 
-import Sidebar from './components/Sidebar';
+import Sidebar from './components/ux/Sidebar';
 import AutomationNode from './components/nodes/AutomationNode';
 import FileUploadNode from './components/nodes/FileUploadNode';
 import FileViewerNode from './components/nodes/FileViewerNode';
 import HDANode from './components/nodes/HDANode';
 import { Stack } from '@mui/joy';
-import { Header } from './components/Header';
+import { Header } from './components/ux/Header';
 import { TabValues } from './enums';
 import { FileEdge } from './components/edges/FileEdge';
 import useCopyPaste from './hooks/useCopyPaste';
 import useUndoRedo from './hooks/useUndoRedo';
-import { UndoRedoPanel } from './components/UndoRedoPanel';
-import { CopyPastePanel } from './components/CopyPastePanel';
+import { UndoRedoPanel } from './components/utils/UndoRedoPanel';
+import { CopyPastePanel } from './components/utils/CopyPastePanel';
 import AssetViewerNode from './components/nodes/AssetViewerNode';
 
 // Main Awful UI component
