+++ sites/awful-ui/src/components/utils/CopyPastePanel.tsx
